- Added support for basic initial/maintenance margin (selected contracts). Tests. - Fixed symbol properties file for IR contracts - Added VIX index to futures symbol nomenclature (futures.cs)

+++ Common/Securities/Option/OptionStrategy.cs
@@ -71,7 +71,7 @@ namespace QuantConnect.Securities.Option
             /// <summary>
             /// Strike price of the leg
             /// </summary>
-            public int Strike { get; set; }
+            public decimal Strike { get; set; }
 
             /// <summary>
             /// Type of order that is to be sent to the market on strategy execution
